tlscontext: add optional dhparam_file field
This is a preparation step for introducing the dhparam-file() TLS option.

dhparam_file can be used to specify Diffie-Hellman parameters generated
via 'openssl dhparam'.

The path should point to a valid DH-params file.
If dhparam_file is not specified, fallback parameters are used.

Parameters are validated with DH_check(), the connection
won't be established if the validation fails.

NOTE: "OpenSSL checks the prime is congruent to 11 when g = 2;
while the IETF's primes are congruent to 23 when g = 2."
For this reason, RFC 3526 parameters can not be used with dhparam-file().

More information about this issue (+ workaround):
https://wiki.openssl.org/index.php/Diffie-Hellman_parameters

Signed-off-by: László Várady <laszlo.varady@balabit.com>