python: remove exception handling from _py_get_method()
All _py_get_method() call sites fall back to some behaviour if the method
does not exist and _py_get_attr_or_null() we are using here already
clears the error state.

This patch removes the exception handling from this function and moves
the error message to debug. It also renames the function to
_py_get_optional_method() to indicate its behavior.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>