grammar: fix bison 3.3 deprecation warnings
This patch removes the use of deprecated bison features, but at this point
depends on an unreleased bison function, namely that api.location.type works
for the C language, which was implemented by my request in:

https://github.com/akimd/bison/commit/9260e5ca4fa62ff971fc52a2dd64474dd60e5990

By merging this branch to master, we would be depending on a future bison
release, probably 3.4.1.  This is added as a dependency in the configure
script, but we should delay inclusion of this patch at the very least until
that bison is released, and maybe even somewhat more until it appears in
distros.

Signed-off-by: Balazs Scheidler <balazs.scheidler@oneidentity.com>