logthrsource: decouple worker from driver
The worker class is now decoupled from the driver, threaded source
implementations must derive from the LogThreadedSourceWorker too.
The new construct_worker() abstract method has to be implemented by
derived classes.

This makes it possible to fix a reload-revert bug and opens up the
possibility of creating multiple worker threads in the future.

Signed-off-by: László Várady <laszlo.varady@balabit.com>