logthrsource: fix reload-revert crash by creating worker in driver_init()
driver_init() and driver_deinit() must be symmetrical, and the worker is
unrefed in driver_deinit() in order to break the circular reference between
the driver and the worker.

Because the worker is unrefed in driver_deinit(), the worker creation has
been moved to driver_init() to avoid crashing in reload.

Signed-off-by: László Várady <laszlo.varady@balabit.com>