afsocket-dest: rename old persist entries
In d29b98c9a261cbfd04dbe6afbb17008fa5108e7c we changed the persist name
from afsocket_dd_connection to afsocket_dd_connections (notice the s)
We also changed the hostname parameters.

There is no API to actually rename an entry in the persist file (We can
do it for runtime use only.), so we actually allocate a new entry with
the new name, and copy the content of the old one.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>