mongodb: use connection pool for better concurrency
https://mongoc.org/libmongoc/current/connection-pooling.html

When the driver is in pooled mode, your program’s operations are unblocked
as soon as monitoring discovers a usable server. For example, if a thread
in your program is waiting to execute an “insert” on the primary, it is
unblocked as soon as the primary is discovered, rather than waiting for
all secondaries to be checked as well.

The pool opens one connection per server for monitoring, and each client
opens its own connection to each server it uses for application operations.

Note: multiple servers can be defined in the uri() option, for example:
"mongodb://hostA,hostB/?replicaSet=my_rs"

Signed-off-by: László Várady <laszlo.varady@protonmail.com>