qdisk: add test for wrapping in disk-buffer, when we never truncate
This is a strange unit test, it was created to help to investigate a bug
and validate its fix. I added this first, so one can checkout this
commit, and see, that it fails with the current code.

The details and the root cause of the bug will be in the next commit's
message.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>