diskq-nonreliable: remove push_head()
push_head() is dead code, it is not used by syslog-ng.

This method can be removed from the LogQueue interface in the future
in favor of the backlog and its rewind mechanism.

Subsequent commits are going to change the internals of the non-reliable
disk-buffer's push() and pop() operations, this was the trigger for this
cleanup.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>