diskq-nonreliable: do not move from qoverflow
This is a preparation step for changing how the non-reliable disk-buffer
moves messages from the qoverflow queue segment.

This and the following 2 commits belong together.

The current move logic serializes and writes all messages from the
qoverflow queue to the disk, which is an expensive operation, especially
in the destination thread. This is done for 2 reasons:
- to retain FIFO ordering relatively easily;
- to wake up sources as soon as possible.

It turned out that the FIFO property can be maintained without doing the
unnecessary serialization/deserialization cycles, and this leads to a
significant performance improvement, which increases the throughput of
the non-reliable disk-buffer and overall does not keep sources in a
suspended state longer than before (unnecessary
serialization/deserialization and disk read/writes under a single
disk-buffer lock produce a blocking, flow-control-like behaviour).

Signed-off-by: László Várady <laszlo.varady@protonmail.com>