stats-cluster-single: store name when initializing with name
Currently we can only store literal strings in StatsCounterGroupInit.

An upcoming feature wants to create a StatsClusterKey with a
heap-allocated string.

This commit copies the string from StatsCounterGroupInit to
StatsCounterGroup, when initializing the Group, and frees it, when
the Group is freed.

The call site only has to make sure, that the heap-allocated string
is available, when the Group gets initialized.

=======================================================================

There are places, where we internally clone the StatsClusterKey, and
use the GroupInit in the cloned Key to initialize the Group.

This clone method currently creates a shallow copy from the GroupInit.

A deep copy will be implemented in the next commit to fix that.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>