nvtable: promote NVEntry->vindirect.type value to NVEntry->type
The old location of the type field was "indirect" value specific,
use the new NVEntry level member to store it and migrate previous
value in deserialization code to the new place, while resetting the
old value to zero.

We can zero out the old location as there is no consumer of the
old type field (I've checked my memories and old sources as well,
there is no API to expose this field). The only scenario where this would
cause a change is:

  1) old syslog-ng serializes message, after being parsed by db-parser()
     which sets the type value
  2) new syslog-ng deserializes the message (from disk buffer for instance),
     and reserializes it again. This will cause the old "type"
     field to become zero.
  3) old syslog-ng deserializes message again, it will find type to be zero.

The only downside would be is that the name-value pair would be
interpreted as a string. I don't know any downstream template functions or
code that would do that today, although it is exactly my intention to
make this possible, thus this series of changes.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>