afsql: replace deprecated MD5 index names
This should NOT cause any compatibility issues as this function will
only be called if
- the destination is oracle SQL
- the table_name and column_name has a combined length > 25

Oracle does not allow 2 identical indexes with different names, but
this will not be a problem either as we only update/create indexes when
the table does not exist or its schema is modified.

The hash function is used as a workaround due to Oracle not allowing
long index names, so the hash itself is truncated to the appropiate
length, hence SHA1 (the shorter the "better").

Signed-off-by: Szilárd Parrag <szilard.parrag@gmail.com>
Signed-off-by: László Várady <laszlo.varady@protonmail.com>