dbld: remove ubuntu-xenial
It does not work with python3.

Its standard support ended in April 2021.
Also the python3 version on it is python3.5, which's support ended
in September 2020.

We have kept xenial until now, because we could use it with python2,
but now, as we started to use python3 everywhere, and it does not work
here, it is time to remove xenial from the supported platforms.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>