python: refactor python-logmsg unit test
`_dict_clone_value()` limits us, to only be able to get python values
as strings.

Also, it extracts the python value to C, which leaves us room for error
(which we happen to have, because we lose the knowledge about whether
it is shown as bytes or str in python)

The only 2 users of `_dict_clone_value()` uses it to check a python
variable's value, so it is easier to add an assert function for that,
which can handle any kinds of types (as you build up a "script", you
can inject any types of variables).

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>