python: introduce backward compatibility mode for 3.x cfg versions
In the previous 2 commits we have introduced 2 breaking changes:
  1. Previously if someone tried to set an NV value with not a
     string/bytes type in a python code, we threw an error.
     With the previous commit we changed the behavior and we store it
     with the correct type.
  2. Previously the python NV accessor function always returned
     string/bytes type. With the previous commit we return the same type
     as it is stored in the `LogMessage`.

The 1. point is not a big deal for backward compatibility reasons,
because we can be more forgiving, as long as we do not fail with
previously working code.

The 2. point is a bigger deal, because older codes expect strings/bytes,
but we can return other types, based on how it is stored in the
`LogMessage`.

To fix these, I have added a compatibility flag to the `PyLogMessage`
class. If the flag is set, we throw an error when setting a non
string/bytes value in python code, and we return string/bytes typed
value when accessing NVs in python code, no matter how it is stored in
the `LogMessage`.

Unfortunately, setting this flag can only be done where we create the
`PyLogMessage` instance in C code, as we only have a reference to the
`GlobalConfig` there. We cannot do that in the python code, because the
constructor there executes a callback (`py_log_message_new_empty()`),
which cannot have any user_data provided. This means we can only set it
in parser and destination, and not in source.
Not setting it in the source is fine, because there, we have always
created an empty `LogMessage`, so there is no risk of accessing NVs with
non bytes/string type.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>