otel: add opentelemetry-proto submodule
We do not have to run autogen.sh in it, and git submodule update --init
is only run by autogen.sh if ivykis is not inited yet.

So if you are using an already submodule inited syslog-ng workspace,
make sure to run git submodule update --init. If you are cloning a new
syslog-ng repo, autogen.sh does the init for you.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>